moved More button to List/List.jsx, added Overall rating to Add button modal
