+++ client/src/Components/ReviewsContainer/Breakdowns/ProductBkdn.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react"
 import Container from "react-bootstrap/Container"
 import Row from "react-bootstrap/Row"
 import Col from "react-bootstrap/Col"
+import Form from "react-bootstrap/Form"
 import useReviews from "../../../contexts/ReviewsContext.jsx"
 
 const scaleStyle = {
