+++ client/src/Components/ReviewsContainer/List/List.jsx
@@ -1,9 +1,23 @@
 import React, { useState } from "react"
+import Container from "react-bootstrap/Container"
+import Row from "react-bootstrap/Row"
+import Col from "react-bootstrap/Col"
+import Button from "react-bootstrap/Button"
 import ReviewList from "./ReviewList.jsx"
-import MoreAddBtns from "./MoreAddBtns.jsx"
+import AddBtn from "./AddBtn.jsx"
+import useReviews from "../../../contexts/ReviewsContext.jsx"
 
 const List = (props) => {
+  const { reviews } = useReviews()
   const [listLength, setListLength] = useState(2)
+
+  const MoreBtn = () => (
+    <Col>
+      <Button className="more" onClick={showMore} style={{ float: "left" }}>
+        More reviews
+      </Button>
+    </Col>
+  )
 
   const showMore = () => {
     setListLength(listLength + 2)
@@ -12,7 +26,12 @@ const List = (props) => {
   return (
     <section className="list">
       <ReviewList sortReviews={props.sortReviews} listLength={listLength} />
-      <MoreAddBtns showMore={showMore} listLength={listLength} />
+      <Container>
+        <Row>
+          {reviews && listLength < reviews.length ? <MoreBtn /> : null}
+          <AddBtn />
+        </Row>
+      </Container>
     </section>
   )
 }
